refactor(http11): ResourceUtil 도입
